Remove jruby support
Last `freddy-jruby` was released Feb 2018. It's almost 2 years older
than `freddy`. As we don't use `freddy-jruby` ourselves and we're not
aware of anyone else using it, there's no reason to continue supporting
it. This commit removes jruby support.
